switched to using Brewfile, made rest of bin scripts idempotent

+++ bin/pip.sh
@@ -1,7 +1,9 @@
 
-curl -o $HOME/downloads/get-pip.py https://bootstrap.pypa.io/get-pip.py
-python $HOME/downloads/get-pip.py
-rm $HOME/downloads/get-pip.py
+if [ ! -f /usr/local/bin/brew ]; then
+	curl -o $HOME/downloads/get-pip.py https://bootstrap.pypa.io/get-pip.py
+	python $HOME/downloads/get-pip.py
+	rm $HOME/downloads/get-pip.py
+fi
 
 if [ ! -d $HOME/.pyenv ]; then
     echo "Installing Python 3.5.2"
